Implemented Mana System and Mana Bar
There is now a working mana system that uses similar methods to the
health system. Mana is now linked to the current player projectile
ability and the telekinesis power. Firing projectiles will drain five
mana a pop until mana is insufficient, in which the player will be
unable to fire until mana regenerates. Telekinesis drains mana slowly
while the player is levitating the object and slightly faster as the
player moves the object. When the player runs out of mana, telekinesis
mode exits as if the player had selected x. There is a blue mana display
bar which depletes and refills.